Add Support for Spark 3
This commit adds support for Spark 3. The APIs used in Spark 2 have
been since deprecated in Spark 3. For that reason, we introduce
sourceSets for Spark 2 and Spark 3. Spark specific code lives in
each sourceSet, and the general implementations are kept in main.

A new profile is added `profiles/scala-2.12-spark3.gradle`, which
brings versions needed for spark3. `rio.yml` is updated to reflect
the new pipelines required for spark3.

A `README.md` and `DEV-README.md` file are added to better support
developers of this project.

Upgrade to Gradle 7.3.3