Add test to re-create issue where: (#190)
1 - partition key filter does not overlap with FourZeroSSTableReader token range so is filtered out
2 - but the spark token range filter does overlap with FourZeroSSTableReader token range so is not filtered out

In this case, we incorrectly do not ignore the sstable because it will not contain the partition in PartitionKeyFilter, we also filter out the PartitionKeyFilter so we do not check the bloom filter or primary Index.db file.

This results in reading the entire SSTable and return all rows even when a PartitionKeyFilter was passed in.